feat: update validator registry pipeline
