+++ node-local-deploy.sh
@@ -19,6 +19,10 @@ export FAUCET_LMR_PAYOUT=200000000 # 2 LMR
 export FAUCET_ETH_PAYOUT=10000000000000000 # 0.01 ETH
 export VALIDATOR_FEE_RATE="0.01"
 export BUILD_FULL_MARKETPLACE=false
+export VALIDATOR_STAKE_MINIMUM=100000000000
+export VALIDATOR_STAKE_REGISTER=1000000000000
+export VALIDATOR_PUNISH_AMOUNT=100000000000
+export VALIDATOR_PUNISH_THRESHOLD=3
 
 # deploy lumerin token
 yarn hardhat run --network localhost --config hardhat-base.config.ts ./scripts/deploy-lumerin.ts
